Add exception

+++ plasmidgenietool_split.py
@@ -25,14 +25,17 @@ if __name__ == '__main__':
     parser = arguments()
     args = parser.parse_args()
     os.path.basename(args.input)
-    z = zipfile.ZipFile(args.input)
-    for f in z.namelist():
-        if f.endswith('_export.csv'):
-            outpath = os.path.dirname(args.output1)
-            z.extract(f,path=outpath)
-            shutil.move(os.path.join(outpath,os.path.basename(f)), args.output1)
-        elif f.endswith('_export_mapping.csv'):
-            outpath = os.path.dirname(args.output2)
-            z.extract(f,path=outpath)
-            shutil.move(os.path.join(outpath,os.path.basename(f)), args.output2)
-    os.unlink(args.input)
+    if os.path.exists(args.input):
+        z = zipfile.ZipFile(args.input)
+        for f in z.namelist():
+            if f.endswith('_export.csv'):
+                outpath = os.path.dirname(args.output1)
+                z.extract(f,path=outpath)
+                shutil.move(os.path.join(outpath,os.path.basename(f)), args.output1)
+            elif f.endswith('_export_mapping.csv'):
+                outpath = os.path.dirname(args.output2)
+                z.extract(f,path=outpath)
+                shutil.move(os.path.join(outpath,os.path.basename(f)), args.output2)
+        os.unlink(args.input)
+    else:
+        print('No output')
